+++ src/pages/api/blog/blogwriter/index.js
@@ -1,9 +1,9 @@
-import multer from 'multer';
-import path from 'path';
-import fs from 'fs';
-import BlogWriter from '@/models/blog/BlogWriter';
-import connectToDatabase from '@/utils/db';
-const uploadDirectory = './uploads/blogwriter';
+import multer from "multer";
+import path from "path";
+import fs from "fs";
+import BlogWriter from "@/models/blog/BlogWriter";
+import connectToDatabase from "@/utils/db";
+const uploadDirectory = path.join(process.cwd(), "uploads/blogwriter");
 if (!fs.existsSync(uploadDirectory)) {
   fs.mkdirSync(uploadDirectory, { recursive: true });
 }
@@ -19,59 +19,60 @@ const storage = multer.diskStorage({
 const upload = multer({ storage });
 
 const apiRoute = async (req, res) => {
-  await connectToDatabase()
-  if (req.method === 'POST'||req.method === 'PUT') {
-    upload.single('file')(req, File, async (err) => {
+  await connectToDatabase();
+  if (req.method === "POST" || req.method === "PUT") {
+    upload.single("file")(req, File, async (err) => {
       if (err instanceof multer.MulterError) {
-        console.error('Multer error:', err);
-        return res.status(500).json({ error: 'File upload failed' });
+        console.error("Multer error:", err);
+        return res.status(500).json({ error: "File upload failed" });
       } else if (err) {
-        console.error('Unknown error during file upload:', err);
-        return res.status(500).json({ error: 'File upload failed' });
+        console.error("Unknown error during file upload:", err);
+        return res.status(500).json({ error: "File upload failed" });
       }
       const { blogwriter } = req.body;
       const fileData = req.file && {
         blogwriter,
-        filename:req.file.filename,
+        filename: req.file.filename,
         path: `/api/uploads/blogwriter/${req.file.filename}`,
-      }
+      };
       try {
-          const file = new BlogWriter(fileData);
-          await file.save();
-          return res.status(200).json({ data: file });
+        const file = new BlogWriter(fileData);
+        await file.save();
+        return res.status(200).json({ data: file });
       } catch (error) {
-        console.error('Error updating or saving file:', error);
-        return res.status(500).json({ error: 'Internal Server Error' });
+        console.error("Error updating or saving file:", error);
+        return res.status(500).json({ error: "Internal Server Error" });
       }
     });
-  } else if (req.method === 'GET') {
+  } else if (req.method === "GET") {
     try {
-     
       const files = await BlogWriter.find({});
       return res.status(200).json({ data: files });
     } catch (error) {
-      console.error('Error fetching files:', error);
-      return res.status(500).json({ error: 'Internal Server Error' });
+      console.error("Error fetching files:", error);
+      return res.status(500).json({ error: "Internal Server Error" });
     }
-  } else if (req.method === 'DELETE') {
+  } else if (req.method === "DELETE") {
     const { id } = req.query;
-    
+
     try {
-      const file = await BlogWriter.findById({_id:id});
+      const file = await BlogWriter.findById({ _id: id });
       if (file) {
-        // console.log("file image id by selected logo ",file);
-        fs.unlinkSync(path.join(uploadDirectory, file.filename));
-        await BlogWriter.findByIdAndDelete({_id:id});
-        return res.status(200).json({ message: 'File removed successfully' });
+        if (fs.existsSync(path.join(uploadDirectory, file.filename))) {
+          // console.log("file image id by selected logo ",file);
+          fs.unlinkSync(path.join(uploadDirectory, file.filename));
+        }
+        await BlogWriter.findByIdAndDelete({ _id: id });
+        return res.status(200).json({ message: "File removed successfully" });
       } else {
-        return res.status(404).json({ error: 'File not found' });
+        return res.status(404).json({ error: "File not found" });
       }
     } catch (error) {
-      console.error('Error deleting file:', error);
-      return res.status(500).json({ error: 'Internal Server Error' });
+      console.error("Error deleting file:", error);
+      return res.status(500).json({ error: "Internal Server Error" });
     }
   } else {
-    res.setHeader('Allow', ['POST', 'GET', 'DELETE']);
+    res.setHeader("Allow", ["POST", "GET", "DELETE"]);
     return res.status(405).json({ error: `Method ${req.method} Not Allowed` });
   }
 };
