+++ src/pages/api/home/index.js
@@ -1,9 +1,9 @@
-import multer from 'multer';
-import path from 'path';
-import fs from 'fs';
-import Banner from '@/models/Home/Bannner';
-import connectToDatabase from '@/utils/db';
-const uploadDirectory = './uploads/banner';
+import multer from "multer";
+import path from "path";
+import fs from "fs";
+import Banner from "@/models/Home/Bannner";
+import connectToDatabase from "@/utils/db";
+const uploadDirectory = path.join(process.cwd(), "uploads/banner");
 if (!fs.existsSync(uploadDirectory)) {
   fs.mkdirSync(uploadDirectory, { recursive: true });
 }
@@ -20,19 +20,19 @@ const storage = multer.diskStorage({
 const upload = multer({ storage });
 
 const apiRoute = async (req, res) => {
-  await connectToDatabase()
+  await connectToDatabase();
 
-  if (req.method === 'POST') {
-    upload.single('file')(req, File, async (err) => {
+  if (req.method === "POST") {
+    upload.single("file")(req, File, async (err) => {
       if (err instanceof multer.MulterError) {
-        console.error('Multer error:', err);
-        return res.status(500).json({ error: 'File upload failed' });
+        console.error("Multer error:", err);
+        return res.status(500).json({ error: "File upload failed" });
       } else if (err) {
-        console.error('Unknown error during file upload:', err);
-        return res.status(500).json({ error: 'File upload failed' });
+        console.error("Unknown error during file upload:", err);
+        return res.status(500).json({ error: "File upload failed" });
       }
 
-      const { title,description } = req.body;
+      const { title, description } = req.body;
       // console.log("title --> ",title)
       // console.log("req.file.filename---->",req.file.filename)
       const fileData = req.file && {
@@ -40,7 +40,7 @@ const apiRoute = async (req, res) => {
         description,
         filename: req.file.filename,
         videoPath: `/api/uploads/banner/${req.file.filename}`,
-      }
+      };
       try {
         const existingFile = await Banner.findOne({});
         if (existingFile) {
@@ -50,8 +50,10 @@ const apiRoute = async (req, res) => {
             { new: true }
           );
           const file = await Banner.findById(existingFile._id);
-          if(file){
-            fs.unlinkSync(path.join(uploadDirectory, file.filename));
+          if (file) {
+            if (fs.existsSync(path.join(uploadDirectory, file.filename))) {
+              fs.unlinkSync(path.join(uploadDirectory, file.filename));
+            }
           }
           return res.status(200).json({ data: updatedFile });
         } else {
@@ -60,36 +62,38 @@ const apiRoute = async (req, res) => {
           return res.status(200).json({ data: file });
         }
       } catch (error) {
-        console.error('Error updating or saving file:', error);
-        return res.status(500).json({ error: 'Internal Server Error' });
+        console.error("Error updating or saving file:", error);
+        return res.status(500).json({ error: "Internal Server Error" });
       }
     });
-  } else if (req.method === 'GET') {
+  } else if (req.method === "GET") {
     try {
       const files = await Banner.find({});
       return res.status(200).json({ data: files });
     } catch (error) {
-      console.error('Error fetching files:', error);
-      return res.status(500).json({ error: 'Internal Server Error' });
+      console.error("Error fetching files:", error);
+      return res.status(500).json({ error: "Internal Server Error" });
     }
-  } else if (req.method === 'DELETE') {
+  } else if (req.method === "DELETE") {
     const { id } = req.query;
-    console.log("file image id by selected logo ",path.join(uploadDirectory));
+    console.log("file image id by selected logo ", path.join(uploadDirectory));
     try {
       const file = await Banner.findById(id);
       if (file) {
-        fs.unlinkSync(path.join(uploadDirectory, file.filename));
+        if (fs.existsSync(path.join(uploadDirectory, file.filename))) {
+          fs.unlinkSync(path.join(uploadDirectory, file.filename));
+        }
         await Banner.findByIdAndDelete(id);
-        return res.status(200).json({ message: 'File removed successfully' });
+        return res.status(200).json({ message: "File removed successfully" });
       } else {
-        return res.status(404).json({ error: 'File not found' });
+        return res.status(404).json({ error: "File not found" });
       }
     } catch (error) {
-      console.error('Error deleting file:', error);
-      return res.status(500).json({ error: 'Internal Server Error' });
+      console.error("Error deleting file:", error);
+      return res.status(500).json({ error: "Internal Server Error" });
     }
   } else {
-    res.setHeader('Allow', ['POST', 'GET', 'DELETE']);
+    res.setHeader("Allow", ["POST", "GET", "DELETE"]);
     return res.status(405).json({ error: `Method ${req.method} Not Allowed` });
   }
 };
